maria: iniciando implementação do email service e adicionando usuario dto

+++ src/main/java/com/cedup/projetolitterae/backend/services/UsuarioService.java
@@ -2,10 +2,14 @@ package com.cedup.projetolitterae.backend.services;
 
 import com.cedup.projetolitterae.backend.dto.ImagemPerfilDto;
 import com.cedup.projetolitterae.backend.dto.LoginUsuarioDto;
+import com.cedup.projetolitterae.backend.dto.UsuarioDto;
+import com.cedup.projetolitterae.backend.emails.MandarEmail;
+import com.cedup.projetolitterae.backend.entities.Biblioteca;
+import com.cedup.projetolitterae.backend.entities.Locacao;
 import com.cedup.projetolitterae.backend.entities.MensagemRetorno;
+import com.cedup.projetolitterae.backend.entities.Resenha;
 import com.cedup.projetolitterae.backend.entities.Usuario;
 import com.cedup.projetolitterae.backend.exceptions.MensagemRetornoException;
-import com.cedup.projetolitterae.backend.repositories.EnderecoRepository;
 import com.cedup.projetolitterae.backend.repositories.UsuarioRepository;
 import com.cedup.projetolitterae.imagens.FileUploadUtil;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -28,6 +32,12 @@ public class UsuarioService{
     private UsuarioRepository repository;
     @Autowired
     private EnderecoService enderecoService;
+    @Autowired
+    private ResenhaService resenhaService;
+    @Autowired
+    private LocacaoService locacaoService;
+    @Autowired
+    private BibliotecaService bibliotecaService;
 
     public Usuario pesquisarPorId(Long id){
         return repository.findById(id).orElse(null);
@@ -42,7 +52,6 @@ public class UsuarioService{
     }
 
     public Usuario login(LoginUsuarioDto login){
-        login.getDataNascimento().setDate(login.getDataNascimento().getDate() + 1);
         Usuario usuario = pesquisarPorId(login.getId());
 
         SimpleDateFormat sdf = new SimpleDateFormat("dd/MM/yyyy");
@@ -86,26 +95,60 @@ public class UsuarioService{
     }
 
     @Transactional
-    public Usuario cadastrarUsuario(Usuario usuario){
+    public Usuario cadastrarUsuario(UsuarioDto usuarioDto){
+        Usuario usuario = fromDto(usuarioDto);
         Random random = new Random();
         usuario.setId(random.nextLong(1000L, 100000L));
-        enderecoService.cadastrarEndereco(usuario.getEnderecoUsuario());
         return repository.save(usuario);
     }
 
     public Usuario alterarUsuario(Usuario novoUsuario){
         Usuario oldUsuario = pesquisarPorId(novoUsuario.getId());
         novoUsuario.getEnderecoUsuario().setId(oldUsuario.getEnderecoUsuario().getId());
-        enderecoService.cadastrarEndereco(novoUsuario.getEnderecoUsuario());
+        enderecoService.alterarEndereco(novoUsuario.getEnderecoUsuario());
         repository.save(novoUsuario);
         return novoUsuario;
     }
 
     public void excluirUsuario(Long id){
         Usuario usuario = pesquisarPorId(id);
-        repository.deleteById(id);
-        enderecoService.excluirEndereco(usuario.getEnderecoUsuario().getId());
-        File imagem = new File(usuario.getImagem());
-        imagem.delete();
+        if(usuario != null){
+            List<Resenha> resenhas = resenhaService.pesquisarResenhaPorIdUsuario(id);
+            resenhas.forEach(x -> resenhaService.excluirResenha(x.getId()));
+
+            List<Locacao> locacoes = locacaoService.pesquisarPorUsuario(id);
+            locacoes.forEach(x -> locacaoService.excluirLocacao(x.getId()));
+
+            repository.deleteById(id);
+            enderecoService.excluirEndereco(usuario.getEnderecoUsuario().getId());
+
+            File imagem = new File(usuario.getImagem());
+            imagem.delete();
+        }
+    }
+
+    private Usuario fromDto(UsuarioDto usuarioDto){
+        Usuario usuario = new Usuario();
+        Biblioteca biblioteca = bibliotecaService.pesquisarPorId(usuarioDto.getIdBiblioteca());
+
+        usuario.setId(usuarioDto.getId());
+        usuario.setCpf(usuarioDto.getCpf());
+        usuario.setNome(usuarioDto.getNome());
+        usuario.setSobrenome(usuarioDto.getSobrenome());
+        usuario.setEnderecoUsuario(enderecoService.cadastrarEndereco(usuarioDto.getEnderecoUsuario()));
+        usuario.setBiblioteca(biblioteca != null ? biblioteca : (Biblioteca) lancaExcecaoDto("id biblioteca"));
+        usuario.setEmail(usuarioDto.getEmail());
+        usuario.setDataNascimento(usuarioDto.getDataNascimento());
+        usuario.setTelefone1(usuarioDto.getTelefone1());
+        usuario.setTelefone2(usuarioDto.getTelefone2());
+        usuario.setImagem(usuarioDto.getImagem());
+        usuario.setTipoPerfil(usuario.getTipoPerfil());
+        usuario.setAtivo(usuarioDto.isAtivo());
+
+        return usuario;
+    }
+
+    private Object lancaExcecaoDto(String campo){
+        throw new MensagemRetornoException(new MensagemRetorno("ERRO", "Campo "+ campo +" não existe."));
     }
 }
